refactor(sqlgen): split oversized dualpath test file and characterization function (#168)
Pure test-scaffolding restructure under coding-standard.md §1 (files ≤500
lines, functions ≤100 lines). No assertion, golden-string, or coverage
changes — every test and subtest is preserved byte-for-byte.

- dualpath_sql_generator_test.go (542 → 290): split by function-under-test.
  The buildPgMainClause cases move to pg_main_clause_test.go; the
  buildDuckClause + ParseDateValue cases move to duck_clause_test.go. The file
  retains the ToDualClauses + classifyPredicate surface.
- predicate_characterization_test.go: TestToDualClauses_Characterization was a
  223-line table-driven function. Extract the 18-case matrix into per-storage-
  class builders (charTextCases / charNumericBoolCases / charTemporalUuidCases
  / charCompositeCases), each <100 lines, and reduce the test body to
  assembling them and running the existing loop. The two other characterization
  tests are untouched.

All 23 test functions preserved; every touched file ≤500 lines and every
function ≤100 lines. go test ./internal/sqlgen/... green, go vet and
golangci-lint (v1.64.8) clean, gofmt clean.
