+++ internal/sqlgen/dualpath_sql_generator_test.go
@@ -255,69 +255,6 @@ func TestToDualClauses_BoundAttributeUnsupportedOperator_ReturnsError(t *testing
 	require.Nil(t, pgArgs)
 }
 
-// Given a DuckDB predicate without metadata, when the clause is built,
-// then value-type fallback inference still produces the expected cast and parameter value.
-func TestBuildDuckClause_NoMetadata_InferTypeAndCast(t *testing.T) {
-	cache := forma.SchemaAttributeCache{}
-
-	// RFC3339 datetime literal with no metadata present.
-	// Use an explicit equals: prefix so the timestamp's ':' characters are not misparsed as an operator separator.
-	cond := &forma.KvCondition{Attr: "unknown_ts", Value: "equals:2020-01-02T03:04:05Z"}
-
-	clause, args, err := buildDuckClause(cond, cache)
-	require.NoError(t, err)
-
-	// Expect a CAST(? AS TIMESTAMP) because the value should be inferred as a datetime
-	require.Contains(t, clause, "CAST(? AS TIMESTAMP)")
-	// Column name should be the attribute name when no metadata is present
-	require.Contains(t, clause, "unknown_ts")
-
-	require.Len(t, args, 1)
-
-	// The argument should be a time.Time equal to the parsed RFC3339 value (in UTC)
-	parsedWant, err := time.Parse(time.RFC3339Nano, "2020-01-02T03:04:05Z")
-	require.NoError(t, err)
-
-	gotTime, ok := args[0].(time.Time)
-	require.True(t, ok, "expected DuckDB arg to be time.Time")
-	require.True(t, gotTime.Equal(parsedWant.UTC()), "expected parsed time to match")
-}
-
-// Given a LIKE-style operator such as starts_with or contains,
-// when a DuckDB clause is built, then the wildcard rewrite is correct
-// and no CAST expression is introduced for the LIKE path.
-func TestBuildDuckClause_LikeOperators_WildcardRewrite_NoCast(t *testing.T) {
-	cache := forma.SchemaAttributeCache{
-		"bio": forma.AttributeMetadata{
-			AttributeID: 40,
-			ValueType:   forma.ValueTypeText,
-			ColumnBinding: &forma.MainColumnBinding{
-				ColumnName: forma.MainColumn("text_10"),
-			},
-		},
-	}
-
-	// starts_with should rewrite to value% and emit a plain LIKE with a single ? placeholder
-	starts := &forma.KvCondition{Attr: "bio", Value: "starts_with:pre"}
-	clause, args, err := buildDuckClause(starts, cache)
-	require.NoError(t, err)
-	require.Contains(t, clause, "bio")
-	require.Contains(t, clause, "LIKE")
-	require.NotContains(t, clause, "CAST(")
-	require.Len(t, args, 1)
-	require.Equal(t, []any{"pre%"}, args)
-
-	// contains should rewrite to %value% and likewise not introduce a CAST
-	contains := &forma.KvCondition{Attr: "bio", Value: "contains:mid"}
-	clause2, args2, err := buildDuckClause(contains, cache)
-	require.NoError(t, err)
-	require.Contains(t, clause2, "bio")
-	require.Contains(t, clause2, "LIKE")
-	require.NotContains(t, clause2, "CAST(")
-	require.Len(t, args2, 1)
-	require.Equal(t, []any{"%mid%"}, args2)
-}
-
 // Given a datetime attribute stored on the main table with Unix-ms encoding,
 // when the Postgres main pushdown is built, then the argument is converted to int64 unix milliseconds.
 func TestToDualClauses_DateMainColumnEncoding_UnixMsArgument(t *testing.T) {
@@ -350,193 +287,4 @@ func TestToDualClauses_DateMainColumnEncoding_UnixMsArgument(t *testing.T) {
 	parsed, err := time.Parse(time.RFC3339Nano, ts)
 	require.NoError(t, err)
 	require.Equal(t, parsed.UnixMilli(), got)
-}
-
-// TestBuildPgMainClause_OR_MixedPushability_SkipsEntireOR verifies that when an OR
-// composite has at least one branch that cannot be pushed to the main table (no column
-// binding), the entire OR is skipped rather than silently emitting a partial OR that
-// would drop rows matched only by the non-pushable branch.
-func TestBuildPgMainClause_OR_MixedPushability_SkipsEntireOR(t *testing.T) {
-	paramIndex := 0
-	cache := forma.SchemaAttributeCache{
-		// branch1 has a column binding → pushable
-		"username": forma.AttributeMetadata{
-			AttributeID: 1,
-			ValueType:   forma.ValueTypeText,
-			ColumnBinding: &forma.MainColumnBinding{
-				ColumnName: forma.MainColumn("text_01"),
-			},
-		},
-		// branch2 has no column binding → NOT pushable
-		"tag": forma.AttributeMetadata{
-			AttributeID: 2,
-			ValueType:   forma.ValueTypeText,
-			// ColumnBinding nil → EAV-only, cannot be pushed to main table
-		},
-	}
-
-	b1 := &forma.KvCondition{Attr: "username", Value: "equals:Alice"}
-	b2 := &forma.KvCondition{Attr: "tag", Value: "equals:admin"}
-	orCond := &forma.CompositeCondition{Logic: forma.LogicOr, Conditions: []forma.Condition{b1, b2}}
-
-	// When: building Postgres main clause for OR with mixed pushability
-	pgClause, pgArgs, err := buildPgMainClause(orCond, cache, &paramIndex)
-	require.NoError(t, err)
-
-	// Then: the entire OR must be skipped (empty clause) to avoid dropping rows that
-	// match only the non-pushable branch. paramIndex must NOT have advanced.
-	require.Equal(t, "", pgClause, "OR with mixed pushability must produce empty main clause")
-	require.Nil(t, pgArgs)
-	require.Equal(t, 0, paramIndex, "paramIndex must not advance when OR is skipped")
-}
-
-// TestBuildPgMainClause_OR_AllPushable_ProducesClause verifies that an OR whose every
-// branch is pushable is correctly emitted as an OR clause.
-func TestBuildPgMainClause_OR_AllPushable_ProducesClause(t *testing.T) {
-	paramIndex := 0
-	cache := forma.SchemaAttributeCache{
-		"first_name": forma.AttributeMetadata{
-			AttributeID: 10,
-			ValueType:   forma.ValueTypeText,
-			ColumnBinding: &forma.MainColumnBinding{
-				ColumnName: forma.MainColumn("text_01"),
-			},
-		},
-		"last_name": forma.AttributeMetadata{
-			AttributeID: 11,
-			ValueType:   forma.ValueTypeText,
-			ColumnBinding: &forma.MainColumnBinding{
-				ColumnName: forma.MainColumn("text_02"),
-			},
-		},
-	}
-
-	b1 := &forma.KvCondition{Attr: "first_name", Value: "equals:Alice"}
-	b2 := &forma.KvCondition{Attr: "last_name", Value: "equals:Smith"}
-	orCond := &forma.CompositeCondition{Logic: forma.LogicOr, Conditions: []forma.Condition{b1, b2}}
-
-	pgClause, pgArgs, err := buildPgMainClause(orCond, cache, &paramIndex)
-	require.NoError(t, err)
-	require.NotEmpty(t, pgClause, "all-pushable OR must produce a clause")
-	require.Contains(t, pgClause, "OR")
-	require.Equal(t, []any{"Alice", "Smith"}, pgArgs)
-	require.Equal(t, 2, paramIndex)
-}
-
-// TestBuildPgMainCompositeClause_EmptyAND_Returns1_1 verifies that an empty AND composite
-// produces "1=1" (vacuously TRUE) from the PG main builder.  This matters when the empty
-// AND is a child of an OR: returning "" would silently drop the TRUE branch and make the
-// OR clause narrower than the logical expression.
-func TestBuildPgMainCompositeClause_EmptyAND_Returns1_1(t *testing.T) {
-	paramIndex := 0
-	cond := &forma.CompositeCondition{Logic: forma.LogicAnd, Conditions: nil}
-	clause, args, err := buildPgMainClause(cond, forma.SchemaAttributeCache{}, &paramIndex)
-	require.NoError(t, err)
-	require.Equal(t, "1=1", clause)
-	require.Empty(t, args)
-	require.Equal(t, 0, paramIndex, "paramIndex must not advance for empty AND")
-}
-
-// TestBuildPgMainCompositeClause_EmptyOR_Returns1_0 verifies that an empty OR composite
-// produces "1=0" (vacuously FALSE) from the PG main builder.
-func TestBuildPgMainCompositeClause_EmptyOR_Returns1_0(t *testing.T) {
-	paramIndex := 0
-	cond := &forma.CompositeCondition{Logic: forma.LogicOr, Conditions: nil}
-	clause, args, err := buildPgMainClause(cond, forma.SchemaAttributeCache{}, &paramIndex)
-	require.NoError(t, err)
-	require.Equal(t, "1=0", clause)
-	require.Empty(t, args)
-	require.Equal(t, 0, paramIndex)
-}
-
-// TestBuildPgMainClause_OR_WithEmptyANDBranch_ProducesNoPrefilter verifies that an OR
-// whose first child is an empty AND (= TRUE) does not silently narrow the prefilter.
-// The result must be "1=1" (or vacuous) because (TRUE OR anything) = TRUE.
-func TestBuildPgMainClause_OR_WithEmptyANDBranch_ProducesNoPrefilter(t *testing.T) {
-	paramIndex := 0
-	cache := forma.SchemaAttributeCache{
-		"username": forma.AttributeMetadata{
-			AttributeID: 1,
-			ValueType:   forma.ValueTypeText,
-			ColumnBinding: &forma.MainColumnBinding{
-				ColumnName: forma.MainColumn("text_01"),
-			},
-		},
-	}
-
-	// (empty AND) OR (username = 'Alice')
-	// Logically = TRUE OR (username='Alice') = TRUE → prefilter should not exclude rows.
-	emptyAnd := &forma.CompositeCondition{Logic: forma.LogicAnd, Conditions: nil}
-	kvCond := &forma.KvCondition{Attr: "username", Value: "equals:Alice"}
-	orCond := &forma.CompositeCondition{
-		Logic:      forma.LogicOr,
-		Conditions: []forma.Condition{emptyAnd, kvCond},
-	}
-
-	pgClause, pgArgs, err := buildPgMainClause(orCond, cache, &paramIndex)
-	require.NoError(t, err)
-
-	// The OR contains a TRUE branch (empty AND → "1=1"), so the result must
-	// not narrow the prefilter below all rows. Check that either:
-	//   (a) the clause is vacuous ("1=1"), or
-	//   (b) the clause contains 1=1 so Postgres evaluates it as always-true.
-	if pgClause != "" {
-		require.Contains(t, pgClause, "1=1",
-			"OR with a TRUE branch must not produce a narrowing prefilter; got %q", pgClause)
-	}
-	// paramIndex should only advance for the username arg (or zero if clause is vacuous).
-	require.LessOrEqual(t, paramIndex, 1)
-	_ = pgArgs
-}
-
-// TestBuildDuckClause_EmptyOR_Returns1_0 verifies that an empty OR composite in the
-// DuckDB clause builder returns "1=0" (matches nothing) rather than "1=1" (matches
-// everything). An empty OR has no accepting branch, so it must reject all rows.
-func TestBuildDuckClause_EmptyOR_Returns1_0(t *testing.T) {
-	cond := &forma.CompositeCondition{Logic: forma.LogicOr, Conditions: nil}
-	clause, args, err := buildDuckClause(cond, forma.SchemaAttributeCache{})
-	require.NoError(t, err)
-	require.Equal(t, "1=0", clause, "empty OR must produce 1=0")
-	require.Empty(t, args)
-}
-
-// TestBuildDuckClause_EmptyAND_Returns1_1 verifies that an empty AND composite returns
-// "1=1" (matches everything), consistent with identity-element semantics for AND.
-func TestBuildDuckClause_EmptyAND_Returns1_1(t *testing.T) {
-	cond := &forma.CompositeCondition{Logic: forma.LogicAnd, Conditions: nil}
-	clause, args, err := buildDuckClause(cond, forma.SchemaAttributeCache{})
-	require.NoError(t, err)
-	require.Equal(t, "1=1", clause, "empty AND must produce 1=1")
-	require.Empty(t, args)
-}
-
-// TestBuildDuckClause_OR_AllNonPushable_Returns1_0 verifies that an OR whose every child
-// produces an empty clause (e.g. unknown attributes) falls back to "1=0" rather than
-// "1=1", so the federated DuckDB query does not match unintended rows.
-func TestBuildDuckClause_OR_AllNonPushable_Returns1_0(t *testing.T) {
-	// Both children reference unknown attributes → buildDuckKvClause returns "unknown_attr op ?"
-	// but the composite collapses to 1=0 when none of them emit a clause.
-	// We test via an explicitly empty OR composite (zero children) to isolate the code path.
-	cond := &forma.CompositeCondition{
-		Logic:      forma.LogicOr,
-		Conditions: []forma.Condition{},
-	}
-	clause, _, err := buildDuckClause(cond, forma.SchemaAttributeCache{})
-	require.NoError(t, err)
-	require.Equal(t, "1=0", clause)
-}
-
-// Migrated from the retired TestConvertDateValueForQuery (#140): ParseDateValue
-// must convert a UnixMs literal to an RFC3339 string for ISO8601-encoded
-// columns, and reject non-date input.
-func TestParseDateValue_ISO8601EncodingFromUnixMs(t *testing.T) {
-	meta := forma.AttributeMetadata{
-		ColumnBinding: &forma.MainColumnBinding{ColumnName: forma.MainColumn("text_02"), Encoding: forma.MainColumnEncodingISO8601},
-	}
-	val, err := ParseDateValue("1700000000000", meta)
-	require.NoError(t, err)
-	require.Equal(t, time.UnixMilli(1700000000000).Format(time.RFC3339), val)
-
-	_, err = ParseDateValue("not-a-date", meta)
-	require.Error(t, err)
 }
